+++ Source/BugsnagBreadcrumb.h
@@ -33,21 +33,61 @@
 #endif
 #endif
 
+typedef NS_ENUM(NSUInteger, BSGBreadcrumbType) {
+    /**
+     *  Any breadcrumb sent via Bugsnag.leaveBreadcrumb()
+     */
+    BSGBreadcrumbTypeManual,
+    /**
+     *  A call to Bugsnag.notify() (internal use only)
+     */
+    BSGBreadcrumbTypeError,
+    /**
+     *  A log message
+     */
+    BSGBreadcrumbTypeLog,
+    /**
+     *  A navigation action, such as pushing a view controller or dismissing an
+     *  alert
+     */
+    BSGBreadcrumbTypeNavigation,
+    /**
+     *  A background process, such performing a database query
+     */
+    BSGBreadcrumbTypeProcess,
+    /**
+     *  A network request
+     */
+    BSGBreadcrumbTypeRequest,
+    /**
+     *  Change in application or view state
+     */
+    BSGBreadcrumbTypeState,
+    /**
+     *  A user event, such as authentication or control events
+     */
+    BSGBreadcrumbTypeUser,
+};
+
+@class BugsnagBreadcrumb;
+
+typedef void(^BSGBreadcrumbConfiguration)(BugsnagBreadcrumb *_Nonnull);
+
 @interface BugsnagBreadcrumb : NSObject
 
-@property(readonly, nullable) NSDate *timestamp;
-@property(readonly, copy, nullable) NSString *message;
+@property(readonly, nonatomic, nullable) NSDate *timestamp;
+@property(readwrite) BSGBreadcrumbType type;
+@property(readwrite, nonatomic, copy, nonnull) NSString *name;
+@property(readwrite, nonatomic, copy, nonnull) NSDictionary *metadata;
 
-- (instancetype _Nullable)initWithMessage:(NSString *_Nullable)message
-                                timestamp:(NSDate *_Nullable)date
-    NS_DESIGNATED_INITIALIZER;
++ (instancetype _Nullable)breadcrumbWithBlock:(BSGBreadcrumbConfiguration _Nonnull)block;
+
 @end
 
 @interface BugsnagBreadcrumbs : NSObject
 
 /**
- * The maximum number of breadcrumbs. Resizable. Must be called from the
- * main thread.
+ * The maximum number of breadcrumbs. Resizable.
  */
 @property(assign, nonatomic, readwrite) NSUInteger capacity;
 
@@ -55,13 +95,20 @@
 @property(assign, readonly) NSUInteger count;
 
 /**
- * Store a new breadcrumb with a provided message. Must be called from the
- * main thread.
+ * Store a new breadcrumb with a provided message.
  */
 - (void)addBreadcrumb:(NSString *_Nonnull)breadcrumbMessage;
 
 /**
- * Clear all stored breadcrumbs. Must be called from the main thread.
+ *  Store a new breadcrumb configured via block. Must be called from the main
+ *  thread
+ *
+ *  @param block configuration block
+ */
+- (void)addBreadcrumbWithBlock:(void(^ _Nonnull)(BugsnagBreadcrumb *_Nonnull))block;
+
+/**
+ * Clear all stored breadcrumbs.
  */
 - (void)clearBreadcrumbs;